+++ src/node/server/index.ts
@@ -16,7 +16,6 @@ import { assetPathPlugin } from './serverPluginAssets'
 import { esbuildPlugin } from './serverPluginEsbuild'
 import { ServerConfig } from '../config'
 import { createServerTransformPlugin } from '../transform'
-import { serviceWorkerPlugin } from './serverPluginServiceWorker'
 import { htmlRewritePlugin } from './serverPluginHtml'
 import { proxyPlugin } from './serverPluginProxy'
 import { createCertificate } from '../utils/createCertificate'
@@ -26,6 +25,7 @@ export { rewriteImports } from './serverPluginModuleRewrite'
 import { sourceMapPlugin, SourceMap } from './serverPluginSourceMap'
 import { webWorkerPlugin } from './serverPluginWebWorker'
 import { wasmPlugin } from './serverPluginWasm'
+import { clientPlugin } from './serverPluginClient'
 
 export type ServerPlugin = (ctx: ServerPluginContext) => void
 
@@ -96,7 +96,7 @@ export function createServer(config: ServerConfig): Server {
     envPlugin,
     moduleResolvePlugin,
     proxyPlugin,
-    serviceWorkerPlugin,
+    clientPlugin,
     hmrPlugin,
     ...(transforms.length || Object.keys(vueCustomBlockTransforms).length
       ? [