+++ src/node/transform.ts
@@ -64,7 +64,11 @@ export function createServerTransformPlugin(
     app.use(async (ctx, next) => {
       await next()
 
-      if (!ctx.body) {
+      if (
+        !ctx.body ||
+        (ctx.type === 'text/html' && !isImportRequest(ctx)) ||
+        resolver.isPublicRequest(ctx.path)
+      ) {
         return
       }